requirement.txt
asgiref==3.8.1
boto3==1.36.2
botocore==1.36.2
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
jmespath==1.0.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.1
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0